requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-mongo==4.0.0
apache-airflow-providers-postgres==5.10.1
apache-airflow-providers-slack==8.7.0
google-analytics-admin==0.23.3
google-analytics-data==0.18.16
google-auth-oauthlib==1.2.1
# nnnn
numpy
pandas
pendulum
psycopg2
pymongo
python-bsonjs
python-dateutil
requests
ShopifyAPI
sqlalchemy
stripe
